Solve Climbing Stairs (cpp)

+++ climbing-stairs/solutions/solution.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     int climbStairs(int n) {
         
-        double result = (pow(((1 + sqrt(5)) / 2), n + 1) - pow(((1 - sqrt(5)) / 2), n + 1)) / sqrt(5);
+        int result = (pow(((1 + sqrt(5)) / 2), n + 1) - pow(((1 - sqrt(5)) / 2), n + 1)) / sqrt(5);
         
-        return (int)round(result);
+        return round(result);
     }
 };
